### 2025-01-14 #7662 @srigas
Thank you for adding our work in pre-review, we look forward to collaborating with you in any way necessary to ensure a flawless review process! :)

The editorialbot failed to generate the paper at its first try, possibly due to a misidentation in the paper.md metadata, which we missed because compilation had no issues in our GitHub Actions automations. Nonetheless, it should work now.

As for possible reviewers, I'm afraid I do not have anyone in particular in mind, as I was only recently introduced to JOSS!
