### 2025-01-28 #7662 @fabian-s
👋  @kazewong @daniel-dodd @jeremiecoullon @KindXiaoming

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
