### 2025-01-28 #7662 @fabian-s
@srigas
while I try to rustle up some reviewers for this, please take a look at the DOI the bot found (see above) and see if it's the right one for your reference.
if so, please add it to your bib-file.
