+++ joss/paper.bib
@@ -140,13 +140,13 @@
   	Year = 2024
 }
 
-@article{Wu:2022,
-  	url = {https://arxiv.org/abs/2207.10289},
-  	Archiveprefix = {arXiv},
-  	Author = {{Wu}, C. and {Zhu}, M. and {Tan}, Q. and {Lu}, L.},
-  	Eprint = {2207.10289},
-  	Journal = {ArXiv e-prints},
-  	Month = jul,
-  	Title = {{A comprehensive study of non-adaptive and residual-based adaptive sampling for physics-informed neural networks}},
-  	Year = 2022
+@article{Wu:2023,
+	title = "{A comprehensive study of non-adaptive and residual-based adaptive sampling for physics-informed neural networks}",
+	journal = {Computer Methods in Applied Mechanics and Engineering},
+	volume = {403},
+	pages = {115671},
+	year = {2023},
+	doi = {10.1016/j.cma.2022.115671},
+	url = {https://www.sciencedirect.com/science/article/pii/S0045782522006260},
+	author = {{Wu}, C. and {Zhu}, M. and {Tan}, Q. and {Kartha}, Y. and {Lu}, L.}
 }
