### 2025-02-11 #7662 @fabian-s
👋 @daniel-dodd @KindXiaoming @canyon289 @ColCarroll

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
