### 2025-02-20 #7830 @editorialbot
[REVIEW]: jaxKAN: A unified JAX framework for Kolmogorov-Arnold Networks
**Submitting author:** @srigas (<a href="http://orcid.org/0009-0009-2352-8709">Spyros Rigas</a>)
**Repository:** https://github.com/srigas/jaxkan
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.5
