### 2025-02-22 367721c
Unpinned dependencies, moved pytest to [doc]
