### 2025-02-22 a141271
Test workflow needs explicit [doc] inclusion now
