+++ .github/workflows/test.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install -e .[all]
+          python -m pip install -e .[all,doc]
 
       - name: Run tests
         run: |
