### 2025-03-03 #7830 @fabian-s
@kazewong @ColCarroll

thank you guys for getting started with the reviews of jaxKAN.

if issues with the code/repo/paper keep you from checking off your checklist items, please open up respective issues in the paper's repo and link them here.

if you expect to need more than 2 weeks from today to finish the initial review, please also let me know here in this thread.
