### 2025-03-03 #7830 @kazewong
@fabian-s I was working on the other JOSS review I mentioned, and I scheduled this week to start more in depth review on this repo. I will start putting down some of the issues for this review in the coming 2-3 days
