### 2025-03-05 srigas/jaxKAN#6 @kazewong
JOSS Review - integration test
https://github.com/openjournals/joss-reviews/issues/7830

The test suite in the repo seems pretty comprehensive. A suggestion I have regarding testing is I believe having some integration tests mirroring the notebooks in the tutorial section of the doc could be useful. For example, one can create a reduced version of the following notebooks in the test to ensure the intended real life use cases are functional

https://jaxkan.readthedocs.io/en/latest/tutorials/Tutorial%201%20-%20Getting%20Started.html
https://jaxkan.readthedocs.io/en/latest/tutorials/Tutorial%203%20-%20PINNs.html

The reason behind this suggestion is `jax` is not stable yet, so having some integration tests should help catching some of the breaking changes upstream in the future.
