### 2025-03-05 38327d5
Added real-world tests following basic tutorials
