+++ tests/test_tutorials.py
@@ -1,0 +1,182 @@
+import pytest
+
+from jaxkan.KAN import KAN
+from jaxkan.utils.PIKAN import sobol_sample, gradf
+
+import jax
+import jax.numpy as jnp
+
+from sklearn.model_selection import train_test_split
+from sklearn.metrics import mean_squared_error
+
+from flax import nnx
+import optax
+
+import numpy as np
+
+
+@pytest.fixture
+def seed():
+    return 42
+
+
+@pytest.fixture
+def req_params():
+    return {'k': 5}
+    
+
+def test_function_fitting(seed, req_params):
+
+    # Define the target function
+    def f(x,y):
+        return x**2 + 2*jnp.exp(y)
+
+    # Generate sample data
+    key = jax.random.PRNGKey(seed)
+    x_key, y_key = jax.random.split(key)
+
+    x1 = jax.random.uniform(x_key, shape=(1000,), minval=-1, maxval=1)
+    x2 = jax.random.uniform(y_key, shape=(1000,), minval=-1, maxval=1)
+
+    y = f(x1, x2).reshape(-1, 1)
+    X = jnp.stack([x1, x2], axis=1)
+
+    # Split the data
+    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=seed)
+    
+    assert X_train.shape == (800, 2), "Incorrect X_train shape after splitting the data"
+    assert X_test.shape == (200, 2), "Incorrect X_tesn shape after splitting the data"
+
+    # Initialize the model
+    n_in, n_out, n_hidden = X_train.shape[1], y_train.shape[1], 6    
+    layer_dims = [n_in, n_hidden, n_hidden, n_out]
+    
+    model = KAN(layer_dims = layer_dims,
+                layer_type = 'cheby',
+                required_parameters = req_params,
+                add_bias = True,
+                rngs = nnx.Rngs(seed)
+               )
+
+    # Initialize optimizer
+    opt_type = optax.adam(learning_rate=0.001)
+    optimizer = nnx.Optimizer(model, opt_type)
+
+    # Define the train loop
+    @nnx.jit
+    def train_step(model, optimizer, X_train, y_train):
+    
+        def loss_fn(model):
+            residual = model(X_train) - y_train
+            loss = jnp.mean((residual)**2)
+    
+            return loss
+    
+        loss, grads = nnx.value_and_grad(loss_fn)(model)
+        optimizer.update(grads)
+    
+        return loss
+
+    # Train the model
+    num_epochs = 2000
+    
+    for epoch in range(num_epochs):
+        # Calculate the loss
+        loss = train_step(model, optimizer, X_train, y_train)
+
+    # Get predictions
+    y_pred = model(X_test)
+    mse = mean_squared_error(y_test, y_pred)
+
+    interv = [0.0003, 0.0004]
+    
+    assert mse < interv[-1] and mse > interv[0], f"Final MSE {mse} does not fall within the range ({interv[0]},{interv[-1]})"
+
+
+def test_pde_solving(seed, req_params):
+
+    # Generate Collocation points for PDE
+    N = 2**12
+    collocs = jnp.array(sobol_sample(np.array([0,-1]), np.array([1,1]), N, seed)) # (4096, 2)
+    
+    # Generate Collocation points for BCs
+    N = 2**6
+    BC1_colloc = jnp.array(sobol_sample(np.array([0,-1]), np.array([0,1]), N)) # (64, 2)
+    BC1_data = - jnp.sin(np.pi*BC1_colloc[:,1]).reshape(-1,1) # (64, 1)
+    BC2_colloc = jnp.array(sobol_sample(np.array([0,-1]), np.array([1,-1]), N)) # (64, 2)
[…]
