### 2025-03-05 a258870
Relaxed the tutorial tests margins
