### 2025-03-05 srigas/jaxKAN#6 @srigas
@kazewong Thank you for the suggestion!

We had already incorporated tests related to the functionalities of the basic functions used in these tutorials (e.g. in the `test_PIKAN.py` tests), but as you pointed out, no tests with "real-world" data and results. To this end, in our latest push (main branch) we have included a `test_tutorials.py` file with tests that correspond to Tutorials 1 and 3 exactly. For these tests, the assertations are connected to the final errors (either of the fit, or of the trained model) instead of the basic functionalities of the inner functions, which are already investigated in other tests.

I won't draft a new release yet, intending to do so after all JOSS-related changes have been applied. If you believe that the issue has been successfully addressed, feel free to close the issue or let me know if you want me to close it :)
