### 2025-03-09 srigas/jaxKAN#6 @kazewong
No worries on the release, indeed it makes sense to only draft a new release only after the entire JOSS publication process. The tests looks good to me. Closing this issue and checking off the testing box
