### 2025-03-18 srigas/jaxKAN#10 @ColCarroll
[joss review] link to github from docs
Also not a blocking issue!

Consider adding a link to the github repo from the first page (or every page!) of RTD. It was only a slight annoyance!

https://github.com/openjournals/joss-reviews/issues/7830
