### 2025-03-18 #7830 @fabian-s
@srigas please have a go at the issues opened by ColCarroll while we wait for @kazewong to finish up their review.
