### 2025-03-18 srigas/jaxKAN#10 @srigas
@ColCarroll  Thank you for your comment! It does not hurt to point to contributions as much as we can, so I included the relevant section explicitly in the GitHub's README file!

I pushed it in the latest commit (passing tests) of the joss branch. I won't draft a new release yet, intending to do so after all JOSS-related changes have been applied. If you believe that the issue has been successfully addressed, feel free to close the issue or let me know if you want me to close it :)
