### 2025-03-18 #7830 @srigas
Dear @fabian-s

I believe we have taken care of all remaining issues in the latest commit pushed to the `joss` branch. Thanks @ColCarroll  for all the suggestions - especially the hint at the code we had missed! :)
