### 2025-03-18 srigas/jaxKAN#11 @kazewong
JOSS Review - Writing suggestion
https://github.com/openjournals/joss-reviews/issues/7830

This is not a blocking issue, but I believe the paper would benefit from some clarification in the **Network methods** section.

Currently the **Network methods** section is under core functionality section. It is unclear to me what exact functionality the **Network Methods** section is providing. It seems to me it has to do with the `grid_updates` function. Is this related to choosing an adaptive grid while deploying the package to solve PDEs? If so, I believe it would be much better to rename the section accordingly.

Another small issue is the full form of PIKAN is never specified. I believe it stands for Physics Informed KAN, right? If so, please write the long form out when mentioning it for the first time.
