### 2025-03-18 srigas/jaxKAN#12 @kazewong
JOSS Review - pinning dependencies.
https://github.com/openjournals/joss-reviews/issues/7830

I highly encourage the authors to pin the version of the packages in their `pyproject.toml`, especially `flax.nnx` does not offer a version 1.0.0 yet, meaning their API can change significantly in the future. Pinning the version of dependencies helps in terms of the longevity of the project.
