### 2025-03-18 srigas/jaxKAN#11 @srigas
@kazewong Thank you for your message!

We added the "Network methods" section under core functionality as the second section after "Layer Selection" because we felt that it is the next step in the order of using the library: after choosing a base layer (described in Layer Selection), we define a KAN (a network of such layers), which has the three methods explained (in a nutshell, init, call, grid_update). Among these methods, the `grid_update` is not relevant only to PDE solving, but to any time of deep learning task, which is why it was included here and not in the "PIKAN utilities" section.

Also, you are correct in that PIKAN stands for Physics-Informed Kolmorogov-Arnold Network. The abbreviation is first mentioned along with the long form in lines 17-18 (final sentence of the Summary section), and a citation is also provided for the paper that first introduced them :)
