### 2025-03-18 srigas/jaxKAN#12 @srigas
@kazewong  Thank you for your message!

The dependencies were originally pinned to certain versions but later unpinned after the discussion in [this](https://github.com/srigas/jaxKAN/issues/4) issue with joss reviewer ColCarroll. We believe that the following configuration presents a middle ground between the two issues:

```
dependencies = [
    "numpy>=2.1.3",
    "scipy>=1.14.1",
    "jax[cpu]>=0.4.35",
    "optax>=0.2.4",
    "flax>=0.10.2",
]
```

If you agree, we can draft a commit and push the changes! :)
