### 2025-03-18 56ef65c
Some ease of life improvements wrt the docs and a minor code fix
