+++ docs/index.rst
@@ -7,6 +7,8 @@ Welcome to **jaxKAN**, a JAX implementation of Kolmogorov-Arnold Networks.
 
 This documentation provides the information necessary to help you get started with jaxKAN. The :ref:`about` page offers further information about the package, while the :ref:`installation` page details how to set it up. A complete :ref:`apiref` is available for exploring the functionality of jaxKAN, and :ref:`tutorials` include ready-to-run example code to guide you through practical applications. For developers interested in contributing, the :ref:`contributing` page outlines how to get involved, and contributions are highly encouraged. Finally, for researchers or enthusiasts who have found jaxKAN useful in their projects, the :ref:`citing` page provides instructions on how to properly reference the package in their work.
 
+The source code for jaxKAN can be found in the `jaxKAN GitHub Repository <https://github.com/srigas/jaxKAN>`_.
+
 .. toctree::
    :maxdepth: 1
    :caption: Contents:
