### 2025-03-18 srigas/jaxKAN#11 @kazewong
Sorry for the oversight for PIKAN's abbreviation, somehow ctrl+f failed me.

Regarding the network methods section, I do believe that section can use a bit of clarification. As it stands now, it is unclear to me how the user would leverage the functionality provide in the "Network methods" section (pardon my ignorance in KANs). As a comparison, the "Layer selection" section is more clear since it states "These are the layers one can **choose** to compose their KANs". For the "Network methods" section, I cannot tell how I would interact with the functionality provided in that section. The three methods provides are init, forward_pass, and grid_updates. The first two are more self explanatory, but the third one is a bit more ambiguous to me. Currently, the choices one would make for the network section seem to be fixed by choosing the layers anyway, i.e. layer size and layer type.

As a more concrete suggestion: can you elaborate on how you imagine the user will use the grid_update methods? Do they have to choose from a list of network methods? Or do they choose the connection between layers?
