### 2025-03-18 srigas/jaxKAN#12 @kazewong
This is fine by me. Please go ahead and make the changes.
