### 2025-03-18 50c3a05
pinned basic dependencies to a minimum version following Issue 12
