+++ pyproject.toml
@@ -29,11 +29,11 @@ license = {file = "LICENSE"}
 requires-python = ">=3.10"
 
 dependencies = [
-    "numpy",
-    "scipy",
-    "jax[cpu]",
-    "optax",
-    "flax",
+    "numpy>=2.1.3",
+    "scipy>=1.14.1",
+    "jax[cpu]>=0.4.35",
+    "optax>=0.2.4",
+    "flax>=0.10.2",
 ]
 
 
