### 2025-03-18 srigas/jaxKAN#12 @kazewong
Tho one thing I would mention is `jax` tends to have breaking changes quite often. I understand pinning the dependencies would make installation harder, but at some point this becomes a problem caused by the dependencies instead of the problem of the packages itself. I would suggest perhaps a more careful approach is to pin within a range, e.g. numpy>=2.1.3, <[version]. This is not a requirement, and I am okay with just pinning the minimum version. But in the future when effort for maintaining this package inevitably reduce, it would be good to considered pinning the version within a range instead of just the minimum version.
