### 2025-03-18 srigas/jaxKAN#11 @srigas
Ah, now I see exactly what you mean @kazewong , sorry for missing this key point!

We have included this (in hindsight, crucial) sentence in the discussed subsection:

> In particular, the `update_grids` method can be called during training with a new grid size as its argument to extend the network's current grid and subsequently adapt the extended grid to the training data.

What do you think?
