### 2025-03-18 srigas/jaxKAN#12 @srigas
Thank you for your comments @kazewong ! At the moment, we have pinned the packages using a minimum version (the changes have been pushed in the latest commit and all tests are passing). In the future we will make sure to use an upper end for the range if necessary! :)
