### 2025-03-18 srigas/jaxKAN#11 @kazewong
Looks good to me, closing this issue now
