### 2025-03-18 #7830 @srigas
@kazewong Thank you for your valuable reviews as well! :)

@fabian-s  We have pushed all joss-related changes in the main branch of jaxKAN and all tests are passing. Additionally, the final version of the paper is built successfully in the `joss` branch of our repo. Let us know if we can help in any other way in the following steps :)
