### 2025-03-18 #7830 @fabian-s
@srigas

At this point could you:

- Make a tagged release of your repo, and list the version tag of the archived version here.
- Archive that release in Zenodo or a similar service (e.g., figshare, an institutional repository, ...)
- Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCIDs.
