+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "jaxkan"
-version = "0.2.0"
+version = "0.2.2"
 description = "A JAX implementation of Kolmogorov-Arnold Networks"
 readme = "README.md"
 keywords = ["JAX", "NNX", "KANs", "Kolmogorov-Arnold", "PIKAN"]
