### 2025-03-18 #7830 @srigas
@fabian-s

We published release `v0.2.4` on GitHub and PyPI and archived it in Zenodo under DOI: [10.5281/zenodo.15045951](https://doi.org/10.5281/zenodo.15045951)

I believe that the metadata are all fine, let me know if there's anything else we can do to assist with the process!
