### 2025-03-18 #7830 @fabian-s
@srigas

please make a release for the `joss` branch (after merging it with `main` to preserve your other changes there) and archive that on Zenodo -- this is the archival deposit of the paper.

Sorry, should have been more clear about that.
