### 2025-03-18 #7830 @srigas
@fabian-s  No worries, we performed the merge and drafted a subsequent release (`v0.2.5`) which is available through PyPI with all tests passing. We also updated the Zenodo archive with the new version, which can be found under DOI: [10.5281/zenodo.15046478](https://doi.org/10.5281/zenodo.15046478)

I will delete the `joss` branch once we're sure that no other changes are necessary there! :)

Let me know if there's anything else I can do.
