### 2025-03-18 #7830 @fabian-s
@kazewong @ColCarroll thank you so much for your timely, diligent, and constructive reviews! JOSS could not exist without brilliant people like you putting in the work.

@srigas thanks for the quick turn-around and congrats on the paper.
Now we need to wait for the section editor to come take a look and take it from here.
