### 2025-03-28 #7830 @srigas
Hello all.

As it's been ten days since the acceptance proposal for jaxKAN, I'm writing this message to make sure that there aren't any actions that you are waiting from our side (i.e., the authors') at this time.

Best,
Spyros.
