### 2025-04-01 4d636fc
Switched from rngs to seed logic
