### 2025-04-01 5d1fb5f
Switched to a layer-grid logic and unified some layer types
