+++ jaxkan/layers/BaseLayer.py
@@ -1,429 +1,0 @@
-from jax import numpy as jnp
-
-from flax import nnx
-
-from ..utils.general import solve_full_lstsq
-
-
-class BaseGrid:
-    """
-    BaseGrid class, corresponding to the grid of the BaseLayer class. It comprises an initialization as well as an update procedure.
-
-    Attributes:
-        n_in (int):
-            Number of layer's incoming nodes.
-        n_out (int):
-            Number of layer's outgoing nodes.
-        k (int):
-            Order of the spline basis functions.
-        G (int):
-            Number of grid intervals.
-        grid_range (tuple):
-            An initial range for the grid's ends, although adaptivity can completely change it.
-        grid_e (float):
-            Parameter that defines if the grids are uniform (grid_e = 1.0) or sample-dependent (grid_e = 0.0). Intermediate values correspond to a linear mixing of the two cases.
-    """
-    
-    def __init__(self, n_in: int = 2, n_out: int = 5, k: int = 3,
-                 G: int = 3, grid_range: tuple = (-1,1), grid_e: float = 0.05
-                ):
-        """
-        Initializes a BaseGrid instance.
-        
-        Args:
-            n_in (int):
-                Number of layer's incoming nodes.
-            n_out (int):
-                Number of layer's outgoing nodes.
-            k (int):
-                Order of the spline basis functions.
-            G (int):
-                Number of grid intervals.
-            grid_range (tuple):
-                An initial range for the grid's ends, although adaptivity can completely change it.
-            grid_e (float):
-                Parameter that defines if the grids are uniform (grid_e = 1.0) or sample-dependent (grid_e = 0.0). Intermediate values correspond to a linear mixing of the two cases.
-            
-        Example:
-            >>> grid_type = BaseGrid(n_in = 2, n_out = 5, k = 3, G = 3, grid_range = (-1,1), grid_e = 0.05)
-            >>> grid = grid_type.item
-        """
-                
-        self.n_in = n_in
-        self.n_out = n_out
-        self.k = k
-        self.G = G
-        self.grid_range = grid_range
-        self.grid_e = grid_e
-
-        # Initialize the grid, which is henceforth callable as .item
-        self.item = self._initialize()
-
-    def _initialize(self):
-        """
-        Create and initialize the grid. Can also be used to reset a grid to the default value.
-
-        Returns:
-            grid (jnp.array):
-                Grid for the BaseLayer, shape (n_in*n_out, G + 2k + 1).
-        
-        Example:
-            >>> grid = BaseGrid(n_in = 2, n_out = 5, k = 3, G = 3, grid_range = (-1,1), grid_e = 0.05)
-            >>> grid.item = grid._initialize()
-        """
-        
-        # Calculate the step size for the knot vector based on its end values
-        h = (self.grid_range[1] - self.grid_range[0]) / self.G
-
-        # Create the initial knot vector and perform augmentation
-        # Now it is expanded from G+1 points to G+1 + 2k points, because k points are appended at each of its ends
-        grid = jnp.arange(-self.k, self.G + self.k + 1, dtype=jnp.float32) * h + self.grid_range[0]
-
-        # Expand for broadcasting - the shape becomes (n_in*n_out, G + 2k + 1), so that the grid
-        # can be passed in all n_in*n_out spline basis functions simultaneously
-        grid = jnp.expand_dims(grid, axis=0)
-        grid = jnp.tile(grid, (self.n_in*self.n_out, 1))
-        
-        return grid
-
-    def update(self, x, G_new):
-        """
[…]
