### 2025-04-01 d004849
k -> D for Chebyshev, Fourier
