+++ jaxkan/layers/Fourier.py
@@ -14,7 +14,7 @@ class FourierLayer(nnx.Module):
             Number of layer's incoming nodes.
         n_out (int):
             Number of layer's outgoing nodes.
-        k (int):
+        D (int):
             Degree of Chebyshev polynomial (1st kind).
         smooth_init (bool):
             Whether to initialize Fourier coefficients with smoothening.
@@ -23,7 +23,7 @@ class FourierLayer(nnx.Module):
     """
     
     def __init__(self,
-                 n_in: int = 2, n_out: int = 5, k: int = 5, smooth_init: bool = True, seed: int = 42
+                 n_in: int = 2, n_out: int = 5, D: int = 5, smooth_init: bool = True, seed: int = 42
                 ):
         """
         Initializes a FourierLayer instance.
@@ -33,7 +33,7 @@ class FourierLayer(nnx.Module):
                 Number of layer's incoming nodes.
             n_out (int):
                 Number of layer's outgoing nodes.
-            k (int):
+            D (int):
                 Degree of Chebyshev polynomial (1st kind).
             smooth_init (bool):
                 Whether to initialize Fourier coefficients with smoothening.
@@ -41,33 +41,33 @@ class FourierLayer(nnx.Module):
                 Random key selection for initializations wherever necessary.
             
         Example:
-            >>> layer = FourierLayer(n_in = 2, n_out = 5, k = 5, smooth_init = True, seed = 42)
+            >>> layer = FourierLayer(n_in = 2, n_out = 5, D = 5, smooth_init = True, seed = 42)
         """
 
         # Setup basic parameters
         self.n_in = n_in
         self.n_out = n_out
-        self.k = k
+        self.D = D
 
         # Setup nnx rngs
         rngs = nnx.Rngs(seed)
         
         # Fourier coefficient normalization
-        norm_factor = jnp.arange(1, self.k + 1) ** 2 if smooth_init else jnp.sqrt(self.k)
+        norm_factor = jnp.arange(1, self.D + 1) ** 2 if smooth_init else jnp.sqrt(self.D)
 
         # Register and initialize the trainable parameters of the layer: c_sin, c_cos
         
         
         # Initialize with σ = 1/sqrt(n_in)
         inits = nnx.initializers.normal(stddev=1.0/jnp.sqrt(self.n_in))(
-                        rngs.params(), (2, self.n_out, self.n_in, self.k), jnp.float32)
+                        rngs.params(), (2, self.n_out, self.n_in, self.D), jnp.float32)
                         
         # Divide by norm_factor, which is either sqrt(k), or the k-dependent smoothening array
         inits /= norm_factor
         
         # Split to sine and cosine terms
-        self.c_cos = nnx.Param(inits[0,:,:,:]) # shape (n_out, n_in, k)
-        self.c_sin = nnx.Param(inits[1,:,:,:]) # shape (n_out, n_in, k)
+        self.c_cos = nnx.Param(inits[0,:,:,:]) # shape (n_out, n_in, D)
+        self.c_sin = nnx.Param(inits[1,:,:,:]) # shape (n_out, n_in, D)
 
     def basis(self, x):
         """
@@ -79,10 +79,10 @@ class FourierLayer(nnx.Module):
 
         Returns:
             c, s (tuple):
-                Cosines, sines applied on inputs, shape (batch, n_in, k).
+                Cosines, sines applied on inputs, shape (batch, n_in, D).
             
         Example:
[…]
