+++ jaxkan/layers/Chebyshev.py
@@ -14,14 +14,14 @@ class ChebyLayer(nnx.Module):
             Number of layer's incoming nodes.
         n_out (int):
             Number of layer's outgoing nodes.
-        k (int):
+        D (int):
             Degree of Chebyshev polynomial (1st kind).
         seed (int):
             Random key selection for initializations wherever necessary.
     """
     
     def __init__(self,
-                 n_in: int = 2, n_out: int = 5, k: int = 5, seed: int = 42
+                 n_in: int = 2, n_out: int = 5, D: int = 5, seed: int = 42
                 ):
         """
         Initializes a ChebyLayer instance.
@@ -31,30 +31,30 @@ class ChebyLayer(nnx.Module):
                 Number of layer's incoming nodes.
             n_out (int):
                 Number of layer's outgoing nodes.
-            k (int):
+            D (int):
                 Degree of Chebyshev polynomial (1st kind).
             seed (int):
                 Random key selection for initializations wherever necessary.
             
         Example:
-            >>> layer = ChebyLayer(n_in = 2, n_out = 5, k = 5, seed = 42)
+            >>> layer = ChebyLayer(n_in = 2, n_out = 5, D = 5, seed = 42)
         """
 
         # Setup basic parameters
         self.n_in = n_in
         self.n_out = n_out
-        self.k = k
+        self.D = D
 
         # Setup nnx rngs
         rngs = nnx.Rngs(seed)
 
         # Register and initialize the trainable parameters of the layer: c_basis, c_act
 
-        # shape (n_out, n_in, k+1)
-        noise_std = 1.0/(self.n_in * (self.k+1))
+        # shape (n_out, n_in, D+1)
+        noise_std = 1.0/(self.n_in * (self.D+1))
         self.c_basis = nnx.Param(
             nnx.initializers.normal(stddev=noise_std)(
-                rngs.params(), (self.n_out, self.n_in, self.k + 1), jnp.float32)
+                rngs.params(), (self.n_out, self.n_in, self.D + 1), jnp.float32)
         )
 
         # shape (n_out, n_in)
@@ -70,10 +70,10 @@ class ChebyLayer(nnx.Module):
 
         Returns:
             x (jnp.array):
-                Chebyshev basis functions applied on inputs, shape (batch, n_in, k+1).
+                Chebyshev basis functions applied on inputs, shape (batch, n_in, D+1).
             
         Example:
-            >>> layer = ChebyLayer(n_in = 2, n_out = 5, k = 5, seed = 42)
+            >>> layer = ChebyLayer(n_in = 2, n_out = 5, D = 5, seed = 42)
             >>>
             >>> key = jax.random.PRNGKey(42)
             >>> x_batch = jax.random.uniform(key, shape=(100, 2), minval=-4.0, maxval=4.0)
@@ -86,51 +86,51 @@ class ChebyLayer(nnx.Module):
         
         x = jnp.expand_dims(x, axis=-1) # (batch, n_in, 1)
 
-        x = jnp.tile(x, (1, 1, self.k + 1)) # (batch, n_in, k+1)
+        x = jnp.tile(x, (1, 1, self.D + 1)) # (batch, n_in, D+1)
 
-        x = jnp.arccos(x) # (batch, n_in, k+1)
+        x = jnp.arccos(x) # (batch, n_in, D+1)
 
-        x *= jnp.arange(self.k+1) # (batch, n_in, k+1)
+        x *= jnp.arange(self.D+1) # (batch, n_in, D+1)
 
-        x = jnp.cos(x) # (batch, n_in, k+1)
+        x = jnp.cos(x) # (batch, n_in, D+1)
 
[…]
