### 2025-04-01 48cb6cb
Added new initialization process
