+++ jaxkan/layers/Spline.py
@@ -1,6 +1,9 @@
-from jax import numpy as jnp
+import jax
+import jax.numpy as jnp
 
 from flax import nnx
+
+from typing import Union
 
 from ..grids.BaseGrid import BaseGrid
 from ..grids.SplineGrid import SplineGrid
@@ -25,20 +28,12 @@ class BaseLayer(nnx.Module):
             An initial range for the grid's ends, although adaptivity can completely change it.
         grid_e (float):
             Parameter that defines if the grids are uniform (grid_e = 1.0) or sample-dependent (grid_e = 0.0). Intermediate values correspond to a linear mixing of the two cases.
-        residual (nn.Module):
+        residual (Union[nnx.Module, None]):
             Function that is applied on samples to calculate residual activation.
-        base_basis (float):
-            std coefficient for initialization of basis weights.
-        base_spline (float):
-            std coefficient for initialization of spline weights.
-        base_res (float):
-            std coefficient for initialization of residual weights.
-        pow_basis (float):
-            Power to raise the 1.0/n_in term for basis weights initialization. 
-        pow_spline (float):
-            Power to raise the 1.0/n_in term for spline weights initialization.
-        pow_res (float):
-            Power to raise the 1.0/n_in term for residual weights initialization.
+        external_weights (bool):
+            Boolean that controls if the trainable weights of shape (n_out, n_in) applied to the splines should be used.
+        init_scheme (Union[dict, None]):
+            Dictionary that defines how the trainable parameters of the layer are initialized.
         seed (int):
             Random key selection for initializations wherever necessary.
     """
@@ -46,13 +41,9 @@ class BaseLayer(nnx.Module):
     def __init__(self,
                  n_in: int = 2, n_out: int = 5, k: int = 3,
                  G: int = 3, grid_range: tuple = (-1,1), grid_e: float = 0.05,
-                 residual: nnx.Module = nnx.silu,
-                 base_basis: float = 1.0,
-                 base_spline: float = 1.0,
-                 base_res: float = 1.0,
-                 pow_basis: float = 0.5,
-                 pow_spline: float = 0.5,
-                 pow_res: float = 0.5,
+                 residual: Union[nnx.Module, None] = nnx.silu,
+                 external_weights: bool = True,
+                 init_scheme: Union[dict, None] = None,
                  seed: int = 42
                 ):
         """
@@ -71,28 +62,19 @@ class BaseLayer(nnx.Module):
                 An initial range for the grid's ends, although adaptivity can completely change it.
             grid_e (float):
                 Parameter that defines if the grids are uniform (grid_e = 1.0) or sample-dependent (grid_e = 0.0). Intermediate values correspond to a linear mixing of the two cases.
-            residual (nn.Module):
+            residual (Union[nnx.Module, None]):
                 Function that is applied on samples to calculate residual activation.
-            base_basis (float):
-                std coefficient for initialization of basis weights.
-            base_spline (float):
-                std coefficient for initialization of spline weights.
-            base_res (float):
-                std coefficient for initialization of residual weights.
-            pow_basis (float):
-                Power to raise the 1.0/n_in term for basis weights initialization. 
-            pow_spline (float):
-                Power to raise the 1.0/n_in term for spline weights initialization.
-            pow_res (float):
-                Power to raise the 1.0/n_in term for residual weights initialization.
+            external_weights (bool):
[…]
