### 2025-04-02 4abee59
Bias is now inhrenet in layer. Will need to implement all these changes for other layers too.
