+++ jaxkan/KAN.py
@@ -19,15 +19,12 @@ class KAN(nnx.Module):
             Type of layer to use (e.g., 'base').
         required_parameters (dict):
             Dictionary containing parameters required for the chosen layer type.
-        add_bias (bool):
-            Boolean that controls wether bias terms are also included during the forward pass or not.
         seed (int):
             Random key selection for initializations wherever necessary.
     """
     
-    def __init__(self,
-                 layer_dims: List[int], layer_type: str = "base", required_parameters: Union[None, dict] = None, 
-                 add_bias: bool = True, seed: int = 42
+    def __init__(self, layer_dims: List[int], layer_type: str = "base",
+                 required_parameters: Union[None, dict] = None, seed: int = 42
                 ):
         """
         Initializes a KAN model.
@@ -46,10 +43,8 @@ class KAN(nnx.Module):
                 
         Example:
             >>> req_params = {'k': 3, 'G': 3, 'grid_range': (-1,1), 'grid_e': 0.05, 'residual': nnx.silu, 'noise_std': 0.1}
-            >>> model = KAN(layer_dims = [2,5,1], layer_type='base', required_parameters=req_params, add_bias = True, seed = 42)
+            >>> model = KAN(layer_dims = [2,5,1], layer_type='base', required_parameters=req_params, seed = 42)
         """
-                
-        self.add_bias = add_bias
         self.layer_type = layer_type.lower()
         
         # Get the corresponding layer class based on layer_type
@@ -68,11 +63,6 @@ class KAN(nnx.Module):
                 for i in range(len(layer_dims) - 1)
             ]
     
-        if self.add_bias:
-            self.biases = [
-                nnx.Param(jnp.zeros((dim,))) for dim in layer_dims[1:]
-            ]
-    
     def update_grids(self, x, G_new):
         """
         Performs the grid update for each layer of the KAN architecture.
@@ -85,7 +75,7 @@ class KAN(nnx.Module):
             
         Example:
             >>> req_params = {'k': 3, 'G': 3, 'grid_range': (-1,1), 'grid_e': 0.05, 'residual': nnx.silu, 'noise_std': 0.1}
-            >>> model = KAN(layer_dims = [2,5,1], layer_type='base', required_parameters=req_params, add_bias = True, seed = 42)
+            >>> model = KAN(layer_dims = [2,5,1], layer_type='base', required_parameters=req_params, seed = 42)
             >>>
             >>> key = jax.random.PRNGKey(42)
             >>> x_batch = jax.random.uniform(key, shape=(100, 2), minval=-4.0, maxval=4.0)
@@ -120,7 +110,7 @@ class KAN(nnx.Module):
             
         Example:
             >>> req_params = {'k': 3, 'G': 3, 'grid_range': (-1,1), 'grid_e': 0.05, 'residual': nnx.silu, 'noise_std': 0.1}
-            >>> model = KAN(layer_dims = [2,5,1], layer_type='base', required_parameters=req_params, add_bias = True, seed = 42)
+            >>> model = KAN(layer_dims = [2,5,1], layer_type='base', required_parameters=req_params, seed = 42)
             >>>
             >>> key = jax.random.PRNGKey(42)
             >>> x_batch = jax.random.uniform(key, shape=(100, 2), minval=-4.0, maxval=4.0)
@@ -131,8 +121,5 @@ class KAN(nnx.Module):
         # Pass through each layer of the KAN
         for i, layer in enumerate(self.layers):
             x = layer(x)
-            
-            if self.add_bias:
-                x += self.biases[i].value
 
         return x
