### 2025-04-02 3cf3c23
Removed some minor bias-related remnants in KAN
