### 2025-04-02 4adc25f
Updated Chebyshev layers with init strategies, flavors, and other changes applied in spline layers.
