+++ jaxkan/layers/Chebyshev.py
@@ -1,13 +1,20 @@
-from jax import numpy as jnp
+import jax
+import jax.numpy as jnp
 
 from flax import nnx
 
+from typing import Union
+
+from ..utils.polynomials import Cb
 from ..utils.general import solve_full_lstsq
         
         
-class ChebyLayer(nnx.Module):
+class ChebyshevLayer(nnx.Module):
     """
-    ChebyLayer class. Corresponds to the Chebyshev version of KANs (ChebyKAN). Ref: https://arxiv.org/pdf/2405.07200
+    ChebyshevLayer class. Corresponds to the Chebyshev version of KANs and comes in three "flavors":
+        "default": the version presented in https://arxiv.org/pdf/2405.07200
+        "modified": the version presented in https://www.sciencedirect.com/science/article/pii/S0045782524005462
+        "exact": uses pre-defined functions for higher efficiency, but cannot scale up to arbitrary degrees
 
     Attributes:
         n_in (int):
@@ -16,15 +23,25 @@ class ChebyLayer(nnx.Module):
             Number of layer's outgoing nodes.
         D (int):
             Degree of Chebyshev polynomial (1st kind).
+        flavor (Union[str, None]):
+            One of "default", "modified", or "exact" - chooses basis implementation.
+        residual (Union[nnx.Module, None]):
+            Function that is applied on samples to calculate residual activation.
+        external_weights (bool):
+            Boolean that controls if the trainable weights (n_out, n_in) should be applied to the activations.
+        init_scheme (Union[dict, None]):
+            Dictionary that defines how the trainable parameters of the layer are initialized.
+        add_bias (bool):
+            Boolean that controls wether bias terms are also included during the forward pass or not.
         seed (int):
             Random key selection for initializations wherever necessary.
     """
     
-    def __init__(self,
-                 n_in: int = 2, n_out: int = 5, D: int = 5, seed: int = 42
-                ):
+    def __init__(self, n_in: int = 2, n_out: int = 5, D: int = 5, flavor: Union[str, None] = None,
+                 residual: Union[nnx.Module, None] = None, external_weights: bool = False,
+                 init_scheme: Union[dict, None] = None, add_bias: bool = True, seed: int = 42):
         """
-        Initializes a ChebyLayer instance.
+        Initializes a ChebyshevLayer instance.
         
         Args:
             n_in (int):
@@ -33,68 +50,207 @@ class ChebyLayer(nnx.Module):
                 Number of layer's outgoing nodes.
             D (int):
                 Degree of Chebyshev polynomial (1st kind).
+            flavor (Union[str, None]):
+                One of "default", "modified", or "exact" - chooses basis implementation.
+            residual (Union[nnx.Module, None]):
+                Function that is applied on samples to calculate residual activation.
+            external_weights (bool):
+                Boolean that controls if the trainable weights (n_out, n_in) should be applied to the activations.
+            init_scheme (Union[dict, None]):
+                Dictionary that defines how the trainable parameters of the layer are initialized.
+            add_bias (bool):
+                Boolean that controls wether bias terms are also included during the forward pass or not.
             seed (int):
                 Random key selection for initializations wherever necessary.
             
         Example:
-            >>> layer = ChebyLayer(n_in = 2, n_out = 5, D = 5, seed = 42)
+            >>> layer = ChebyshevLayer(n_in = 2, n_out = 5, D = 5, flavor = "default", 
[…]
