### 2025-04-02 d046a4e
Added support for Legendre layers
