+++ jaxkan/layers/Legendre.py
@@ -5,15 +5,14 @@ from flax import nnx
 
 from typing import Union
 
-from ..utils.polynomials import Cb
+from ..utils.polynomials import Lg
 from ..utils.general import solve_full_lstsq
         
         
-class ChebyshevLayer(nnx.Module):
+class LegendreLayer(nnx.Module):
     """
-    ChebyshevLayer class. Corresponds to the Chebyshev version of KANs and comes in three "flavors":
-        "default": the version presented in https://arxiv.org/pdf/2405.07200
-        "modified": the version presented in https://www.sciencedirect.com/science/article/pii/S0045782524005462
+    LegendreLayer class. Corresponds to the Legendre version of KANs and comes in two "flavors":
+        "default": uses the recursion formula to calculate polynomials up to arbitrary degree
         "exact": uses pre-defined functions for higher efficiency, but cannot scale up to arbitrary degrees
 
     Attributes:
@@ -22,7 +21,7 @@ class ChebyshevLayer(nnx.Module):
         n_out (int):
             Number of layer's outgoing nodes.
         D (int):
-            Degree of Chebyshev polynomial (1st kind).
+            Degree of Legendre polynomial.
         flavor (Union[str, None]):
             One of "default", "modified", or "exact" - chooses basis implementation.
         residual (Union[nnx.Module, None]):
@@ -41,7 +40,7 @@ class ChebyshevLayer(nnx.Module):
                  residual: Union[nnx.Module, None] = None, external_weights: bool = False,
                  init_scheme: Union[dict, None] = None, add_bias: bool = True, seed: int = 42):
         """
-        Initializes a ChebyshevLayer instance.
+        Initializes a LegendreLayer instance.
         
         Args:
             n_in (int):
@@ -49,7 +48,7 @@ class ChebyshevLayer(nnx.Module):
             n_out (int):
                 Number of layer's outgoing nodes.
             D (int):
-                Degree of Chebyshev polynomial (1st kind).
+                Degree of Legendre polynomial.
             flavor (Union[str, None]):
                 One of "default", "modified", or "exact" - chooses basis implementation.
             residual (Union[nnx.Module, None]):
@@ -64,14 +63,14 @@ class ChebyshevLayer(nnx.Module):
                 Random key selection for initializations wherever necessary.
             
         Example:
-            >>> layer = ChebyshevLayer(n_in = 2, n_out = 5, D = 5, flavor = "default", 
-            >>>                        residual = None, external_weights = False, init_scheme = None,
-            >>>                        add_bias = True, seed = 42)
+            >>> layer = LegendreLayer(n_in = 2, n_out = 5, D = 5, flavor = "default", 
+            >>>                       residual = None, external_weights = False, init_scheme = None,
+            >>>                       add_bias = True, seed = 42)
         """
         if flavor is None:
             flavor = "default"
         elif flavor == "exact":
-            max_deg = max(list(Cb.keys()))
+            max_deg = max(list(Lg.keys()))
             if D > max_deg:
                 raise ValueError(f"For method 'exact', the maximum degree cannot exceed {max_deg}.")
 
@@ -111,20 +110,20 @@ class ChebyshevLayer(nnx.Module):
 
     def basis(self, x):
         """
-        Based on the degree and flavor, the values of the Chebyshev basis functions are calculated on the input.
+        Based on the degree and flavor, the values of the Legendre basis functions are calculated on the input.
 
         Args:
             x (jnp.array):
                 Inputs, shape (batch, n_in).
 
         Returns:
-            cheb (jnp.array):
[…]
