### 2025-04-02 a9bf879
Modified Fourier layer; did not add all spline/polynomial methods because of its distinct nature.
