+++ jaxkan/layers/RBF.py
@@ -1,0 +1,336 @@
+import jax
+import jax.numpy as jnp
+
+from flax import nnx
+
+from typing import Union
+
+from ..grids.RBFGrid import RBFGrid
+from ..utils.general import solve_full_lstsq
+        
+        
+class RBFLayer(nnx.Module):
+    """
+    RBFLayer class. Corresponds to the RBF version of KANs using different kernels.
+
+    Attributes:
+        n_in (int):
+            Number of layer's incoming nodes.
+        n_out (int):
+            Number of layer's outgoing nodes.
+        D (int):
+            Number of basis functions.
+        kernel (Union[str, None]):
+            Kernel to be used for the RBFs.
+        grid_range (tuple):
+            An initial range for the grid's ends, although adaptivity can completely change it.
+        grid_e (float):
+            Parameter that defines if the grids are uniform (grid_e = 1.0) or sample-dependent (grid_e = 0.0). Intermediate values correspond to a linear mixing of the two cases.
+        residual (Union[nnx.Module, None]):
+            Function that is applied on samples to calculate residual activation.
+        external_weights (bool):
+            Boolean that controls if the trainable weights (n_out, n_in) should be applied to the activations.
+        init_scheme (Union[dict, None]):
+            Dictionary that defines how the trainable parameters of the layer are initialized.
+        add_bias (bool):
+            Boolean that controls wether bias terms are also included during the forward pass or not.
+        seed (int):
+            Random key selection for initializations wherever necessary.
+    """
+    
+    def __init__(self, n_in: int = 2, n_out: int = 5, D: int = 5, kernel: Union[str, None] = None,
+                 grid_range: tuple = (-2.0, 2.0), grid_e: float = 1.0,
+                 residual: Union[nnx.Module, None] = None, external_weights: bool = False,
+                 init_scheme: Union[dict, None] = None, add_bias: bool = True, seed: int = 42):
+        """
+        Initializes a RBFLayer instance.
+        
+        Args:
+            n_in (int):
+                Number of layer's incoming nodes.
+            n_out (int):
+                Number of layer's outgoing nodes.
+            D (int):
+                Number of basis functions.
+            kernel (Union[str, None]):
+                Kernel to be used for the RBFs.
+            grid_range (tuple):
+                An initial range for the grid's ends, although adaptivity can completely change it.
+            grid_e (float):
+                Parameter that defines if the grids are uniform (grid_e = 1.0) or sample-dependent (grid_e = 0.0). Intermediate values correspond to a linear mixing of the two cases.
+            residual (Union[nnx.Module, None]):
+                Function that is applied on samples to calculate residual activation.
+            external_weights (bool):
+                Boolean that controls if the trainable weights (n_out, n_in) should be applied to the activations.
+            init_scheme (Union[dict, None]):
+                Dictionary that defines how the trainable parameters of the layer are initialized.
+            add_bias (bool):
+                Boolean that controls wether bias terms are also included during the forward pass or not.
+            seed (int):
+                Random key selection for initializations wherever necessary.
+            
+        Example:
+            >>> layer = RBFLayer(n_in = 2, n_out = 5, D = 4, kernel = "gaussian",
+            >>>                  grid_range = (-2.0, 2.0), grid_e = 1.0, residual = None,
+            >>>                  external_weights = False, init_scheme = None, add_bias = True,
[…]
