+++ jaxkan/grids/BaseGrid.py
@@ -55,6 +55,7 @@ class BaseGrid:
         # Initialize the grid, which is henceforth callable as .item
         self.item = self._initialize()
 
+    
     def _initialize(self):
         """
         Create and initialize the grid. Can also be used to reset a grid to the default value.
@@ -82,6 +83,7 @@ class BaseGrid:
         
         return grid
 
+    
     def update(self, x, G_new):
         """
         Update the grid based on input data and new grid size.
@@ -93,8 +95,8 @@ class BaseGrid:
                 New grid size in terms of intervals.
             
         Example:
-            >>> key = jax.random.PRNGKey(42)
-            >>> x_batch = jax.random.uniform(key, shape=(100, 2), minval=-4.0, maxval=4.0)
+            >>> key = jax.random.key(42)
+            >>> x_batch = jax.random.uniform(key, shape=(100, 2), minval=-1.0, maxval=1.0)
             >>>
             >>> grid = BaseGrid(n_in = 2, n_out = 5, k = 3, G = 3, grid_range = (-1,1), grid_e = 0.05)
             >>> grid.update(x=x_batch, G_new=5)
@@ -147,3 +149,4 @@ class BaseGrid:
         # Update the grid value and size
         self.item = grid
         self.G = G_new
+        
