### 2025-04-02 b431c8e
Added Sine Layer.
