+++ jaxkan/layers/Sine.py
@@ -1,0 +1,326 @@
+import jax
+import jax.numpy as jnp
+
+from flax import nnx
+
+from typing import Union
+
+from ..utils.general import solve_full_lstsq
+        
+        
+class SineLayer(nnx.Module):
+    """
+    SineLayer class, inspired from the sine basis functions introduced in https://arxiv.org/pdf/2410.01990
+
+    Attributes:
+        n_in (int):
+            Number of layer's incoming nodes.
+        n_out (int):
+            Number of layer's outgoing nodes.
+        D (int):
+            Number of basis functions.
+        residual (Union[nnx.Module, None]):
+            Function that is applied on samples to calculate residual activation.
+        external_weights (bool):
+            Boolean that controls if the trainable weights (n_out, n_in) should be applied to the activations.
+        init_scheme (Union[dict, None]):
+            Dictionary that defines how the trainable parameters of the layer are initialized.
+        add_bias (bool):
+            Boolean that controls wether bias terms are also included during the forward pass or not.
+        seed (int):
+            Random key selection for initializations wherever necessary.
+    """
+    
+    def __init__(self, n_in: int = 2, n_out: int = 5, D: int = 5,
+                 residual: Union[nnx.Module, None] = None, external_weights: bool = False,
+                 init_scheme: Union[dict, None] = None, add_bias: bool = True, seed: int = 42):
+        """
+        Initializes a SineLayer instance.
+        
+        Args:
+            n_in (int):
+                Number of layer's incoming nodes.
+            n_out (int):
+                Number of layer's outgoing nodes.
+            D (int):
+                Number of basis functions.
+            residual (Union[nnx.Module, None]):
+                Function that is applied on samples to calculate residual activation.
+            external_weights (bool):
+                Boolean that controls if the trainable weights (n_out, n_in) should be applied to the activations.
+            init_scheme (Union[dict, None]):
+                Dictionary that defines how the trainable parameters of the layer are initialized.
+            add_bias (bool):
+                Boolean that controls wether bias terms are also included during the forward pass or not.
+            seed (int):
+                Random key selection for initializations wherever necessary.
+            
+        Example:
+            >>> layer = SineLayer(n_in = 2, n_out = 5, D = 5, residual = None, external_weights = False,
+            >>>                   init_scheme = None, add_bias = True, seed = 42)
+        """
+
+        # Setup basic parameters
+        self.n_in = n_in
+        self.n_out = n_out
+        self.D = D
+        self.residual = residual
+
+        # Setup nnx rngs
+        self.rngs = nnx.Rngs(seed)
+
+        # Add bias
+        if add_bias == True:
+            self.bias = nnx.Param(jnp.zeros((n_out,)))
+        else:
+            self.bias = None
+
+        # Initialize omegas from N(0,1) - shape (D, 1)
+        self.omega = nnx.Param(
+            nnx.initializers.normal(stddev = 1.0)(
+                self.rngs.params(), (D, 1), jnp.float32)
+        )
+
+        # Initialize phases at 0 - shape (D, 1)
+        self.phase = nnx.Param(jnp.zeros((D, 1)))
+
+        # If external_weights == True, we initialize weights for the activation functions equal to unity
+        if external_weights == True:
+            self.c_ext = nnx.Param(
+                nnx.initializers.ones(
+                    self.rngs.params(), (self.n_out, self.n_in), jnp.float32)
+            )
[…]
