### 2025-04-07 #7830 @srigas
@crvernon  Hello there! I hate to bother you, but I saw some progress in other accepted submissions of this track the past week and was wondering if there are any updates on our front :)

I wouldn't be asking, but there is a paper currently accepted for publication in another journal, where jaxKAN is cited, and it would be ideal if our submission could proceed, so that the proper DOI can be assigned for citation reasons.

Thank you for understanding!
